Update the iron-list internal collection when mutating items.
iron-list uses the collection to map items to indexes etc., but currently
it's immutable to changes. This causes problems when filtering and the items
change.

Another solution to this problem would be to always replace the _cachedItems
with a totally new object when filtering, but that would be a really heavy
operation with large dataset sizes.