Use focus to determine if a input/button has been clicked.
When Shadow DOM is enabled, we can't reliably detect if the <div class="item">
has been focused, so moved the tabindex to data-table-row elements which rely
in the light DOM.

As a trade-off the iron-list arrow navigation doesn't work anymore, but we
need to anyway come up with a custom navigation later on.

This also fixes the :focus:after styles to work on Shadow DOM again.